+++ tests/Company/CompanyServiceTest.php
@@ -18,6 +18,7 @@ use Wizaplace\SDK\Company\UnauthenticatedCompanyRegistration;
 use Wizaplace\SDK\Exception\CompanyNotFound;
 use Wizaplace\SDK\Tests\ApiTestCase;
 use Wizaplace\SDK\Tests\File\Mock;
+use Wizaplace\SDK\User\User;
 use Wizaplace\SDK\User\UserType;
 
 /**
@@ -211,7 +212,7 @@ final class CompanyServiceTest extends ApiTestCase
         $this->assertSame('40 rue Laure Diebold', $company->getAddress());
         $this->assertSame('Paris', $company->getCity());
         $this->assertSame('FR', $company->getCountry());
-        $this->assertSame('01 02 03 04 05', $company->getPhoneNumber());
+        $this->assertSame('0987654321', $company->getPhoneNumber());
         $this->assertSame('the-world-company-inc.', $company->getSlug());
     }
 
@@ -228,7 +229,7 @@ final class CompanyServiceTest extends ApiTestCase
         $this->assertSame('40 rue Laure Diebold', $company->getAddress());
         $this->assertSame('Paris', $company->getCity());
         $this->assertSame('FR', $company->getCountry());
-        $this->assertSame('01 02 03 04 05', $company->getPhoneNumber());
+        $this->assertSame('0987654321', $company->getPhoneNumber());
         $this->assertSame('the-world-company-inc.', $company->getSlug());
     }
 
@@ -236,6 +237,28 @@ final class CompanyServiceTest extends ApiTestCase
     {
         $this->expectException(ClientException::class);
         $this->buildUserCompanyService('vendor@world-company.com', 'password-vendor')->getCompany(1);
+    }
+
+    public function testAddACompanyImageAndDeleteIt()
+    {
+        $service = $this->buildUserCompanyService('vendor@wizaplace.com', 'password');
+        $companyId = $service->getCompany(2)->getId();
+
+        $file = $this->mockUploadedFile('favicon.png');
+
+        $imageId = $service->updateCompanyImage($companyId, [
+            'name'     => "file",
+            'contents' => $file->getStream(),
+            'filename' => $file->getClientFilename(),
+        ]);
+        $this->assertGreaterThan(0, $imageId);
+
+        $companyImageId = $service->getCompanyImageId($companyId);
+        $this->assertEquals(13, $companyImageId);
+
+        $result = $service->deleteCompanyImage($companyId, $companyImageId);
+        $this->assertEquals(true, $result["success"]);
+        $this->assertEquals("Image ".$imageId." successfully deleted", $result["message"]);
     }
 
     public function testGettingAListOfDivisionsCountriesCode(): void